requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyarrow>=12.0.0
requests>=2.31.0
python-dotenv>=1.0.0
praw>=7.7.0
google-api-python-client>=2.100.0
plotly>=5.15.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2023.3
paho-mqtt>=1.6.1

